Add extra wrapper functions around get. Allow `Into<T>` for default
The Into<T> addition allows you to use a &str as the default value for
a call that returns a String. This meant that calls to `require`
with `None` for the default argument, Rust would no longer be able
to infer all the types, thus requiring a 2-argument turbofish where
the second argument wasn't used anyway.

As a result, I changed `require` to not take a default argument,
and added additional `optional` and `with_default` functions to
handle the other two cases and avoid any inconvenient type inference
issues.